Raise tasks useState definition from ListTasks to ListSection
