+++ tristan/ding-wall/src/components/ListSection.js
@@ -1,17 +1,33 @@
-import React from 'react';
+import React, { useState } from 'react';
 import './ListSection.css';
 import ListTasks from './ListTasks';
-// import InputForm from './InputForm';
+import InputForm from './InputForm';
 
-/* This file seems redundant now, since InputForm is now a child of 
-ListTasks, though that was not what I set out to do.
+/* 
+To do: add ability to edit (complete? i.e. strikethrough) and remove tasks, 
+or at least remove.
+
+Note that I relied heavily on the tutorial at
+https://www.pusher.com/tutorials/todo-app-react-hooks
 */
+
 function ListSection() {
+
+  const initialTasks = [{text: "buy lumber"},
+                        {text: "fix the house"},
+                        {text: "move in"}];
+
+  const [tasks, setTasks] = useState(initialTasks);  
+
+  const addTask = text => {
+    setTasks([...tasks, { text }]);
+  };
+
   return (
     <section className="list-container">
       <h2>The list:</h2>
-      {/* <InputForm /> */}
-      <ListTasks />
+      <InputForm addTask={addTask} />
+      <ListTasks tasks={tasks}/>
     </section>
   );
 }
